refactor: RacingCarGame 로직 리팩토링 (메서드 추출)
